refactor: split tui/render/detail por seção
